requirements.txt
pyTelegramBotAPI
pytest
jproperties
flake8
matplotlib
coverage
pytest-mock
python-telegram-bot
python-telegram-bot-calendar
currencyapicom

Flask
Flask-PyMongo
Flask-JWT-Extended
Flask-Mail
pymongo
werkzeug
python-dotenv
